fix: address implicit list flattening
If required-directories only has one element then it was possible for
entries to be flattened from list<list<string>> to just list<string>.
The aforementioned behavior is reasonable; my initial approach was
fundamentally flawed from the beginning.

The solution to all of this is quite simple; all we have to do is use a
record. (I am somewhat confused about why the record is transposed
sometimes, but it works regardless!).